refactor(upstream): split session parse into quota, freebucks, standing, availability
