+++ backend/internal/upstream/session_parse.go
@@ -7,11 +7,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"regexp"
-	"sort"
-	"strconv"
-	"strings"
-	"sync"
 	"time"
 )
 
@@ -85,486 +80,10 @@ type SessionState struct {
 	UpgradeHint *SessionUpgradeHint
 }
 
-// FreeWindowsInfo mirrors upstream FreebuffFreeWindowsInfo (free-tier
-// session pool windows; display-only).
-type FreeWindowsInfo struct {
-	DayUsed      float64   `json:"dayUsed"`
-	DayLimit     float64   `json:"dayLimit"`
-	WeekUsed     float64   `json:"weekUsed"`
-	WeekLimit    float64   `json:"weekLimit"`
-	MonthUsed    float64   `json:"monthUsed"`
-	MonthLimit   float64   `json:"monthLimit"`
-	DayResetAt   time.Time `json:"dayResetAt"`
-	MonthResetAt time.Time `json:"monthResetAt"`
-}
-
-// SessionUpgradeHint mirrors the upstream upgradeHint wire shape
-// (common/src/types/freebuff-session.ts:323-326).
-type SessionUpgradeHint struct {
-	URL     string `json:"url"`
-	Message string `json:"message"`
-}
-
-// SubscriptionInfo mirrors upstream FreebuffSubscriptionUsage (subscriber
-// usage rings + provider spend; rollout audience only).
-type SubscriptionInfo struct {
-	DayUsed            float64   `json:"dayUsed"`
-	DayLimit           float64   `json:"dayLimit"`
-	FiveDayUsed        float64   `json:"fiveDayUsed"`
-	FiveDayLimit       float64   `json:"fiveDayLimit"`
-	MonthUsed          float64   `json:"monthUsed"`
-	MonthLimit         float64   `json:"monthLimit"`
-	DayPremiumUsed     float64   `json:"dayPremiumUsed"`
-	DayPremiumLimit    float64   `json:"dayPremiumLimit"`
-	DayResetAt         time.Time `json:"dayResetAt"`
-	PeriodEndsAt       time.Time `json:"periodEndsAt"`
-	MonthSpendUsd      float64   `json:"monthSpendUsd"`
-	MonthSpendLimitUsd float64   `json:"monthSpendLimitUsd"`
-	FreeDayUsed        *float64  `json:"freeDayUsed,omitempty"`
-	FreeDayLimit       *float64  `json:"freeDayLimit,omitempty"`
-}
-
-// SessionReferral mirrors the upstream FreebuffReferralInfo wire block.
-type SessionReferral struct {
-	Code                    string
-	ReferrerName            string
-	QualifiedCount          int
-	WeeklySessionsRemaining int
-	ResetAt                 time.Time
-	GithubLinked            bool
-}
-
-// FreebucksWindow is one window of a Freebucks allowance.
-type FreebucksWindow struct {
-	Limit     float64   `json:"limit"`
-	Spent     float64   `json:"spent"`
-	Remaining float64   `json:"remaining"`
-	ResetAt   time.Time `json:"resetAt"`
-}
-
-// FreebucksWallet is the never-expiring Freebucks store (issue #321 wire
-// drift): plan bonuses land here (monthlyBonus at nextBonusAt; 0 on free).
-type FreebucksWallet struct {
-	Balance      float64   `json:"balance"`
-	MonthlyBonus float64   `json:"monthlyBonus"`
-	NextBonusAt  time.Time `json:"nextBonusAt,omitempty"`
-}
-
-// FreebucksSpendCeiling is the settled-USD daily spend cap for the account's
-// tier (issue #321 wire drift).
-type FreebucksSpendCeiling struct {
-	LimitUsd float64   `json:"limitUsd"`
-	ResetAt  time.Time `json:"resetAt,omitempty"`
-}
-
-// FreebucksMonthlyAllowance is the monthly dollar allowance (wire drift
-// 2026-09-04, issue #330): provider spend for the period at which fresh
-// sessions stop. Pointer in FreebucksInfo: absent on servers that predate
-// it, and clients must render nothing (not a zero) in that case.
-type FreebucksMonthlyAllowance struct {
-	LimitUsd     float64   `json:"limitUsd"`
-	SpentUsd     float64   `json:"spentUsd"`
-	RemainingUsd float64   `json:"remainingUsd"`
-	ResetAt      time.Time `json:"resetAt"`
-}
-
-// FreebucksPriceChange is one server-announced scheduled repricing (wire
-// drift 2026-09-05, issue #350): applied only once due, never to admitted
-// sessions, and only to models already on the meter.
-type FreebucksPriceChange struct {
-	At      string  `json:"at"`
-	ModelID string  `json:"modelId"`
-	Price   float64 `json:"price"`
-	Tagline string  `json:"tagline"`
-}
-
-// FreebucksInfo is the caller's Freebucks position (issue #232, shape
-// issue #321): spendable balance (= daily.remaining + wallet.balance) +
-// the daily pool + the never-expiring wallet + the USD spend ceiling +
-// the plan id ("" when the account is on the free allowance) +
-// the server-authorized quota exemption + per-model prices with their
-// display copy + the announced repricing schedule (issue #350).
-type FreebucksInfo struct {
-	Balance float64                    `json:"balance"`
-	Daily   FreebucksWindow            `json:"daily"`
-	Wallet  FreebucksWallet            `json:"wallet"`
-	Spend   FreebucksSpendCeiling      `json:"spend"`
-	Monthly *FreebucksMonthlyAllowance `json:"monthly,omitempty"`
-	PlanID  string                     `json:"planId,omitempty"`
-	// QuotaExempt: new sessions stay usable at zero balance (server-sent;
-	// the meter's canStart is exempt || balance >= price).
-	QuotaExempt bool               `json:"quotaExempt,omitempty"`
-	Prices      map[string]float64 `json:"prices"`
-	// PriceNotices overrides the static model tagline with price-resolved
-	// copy (mirrors taglineFor in freebuff-model-selector.tsx).
-	PriceNotices map[string]string      `json:"priceNotices,omitempty"`
-	PriceChanges []FreebucksPriceChange `json:"priceChanges,omitempty"`
-}
-
-// ApplyFreebucksPriceChanges applies the server's announced repricing
-// schedule to already-parsed info (mirrors applyFreebucksPriceChanges in
-// freebuff-price-changes.ts): due changes (at <= now) apply in chronological
-// order, reprice only models already on the meter, refresh their notice
-// copy, and are consumed; future changes are kept for the next call.
-func ApplyFreebucksPriceChanges(fb *FreebucksInfo, now time.Time) {
-	if fb == nil || len(fb.PriceChanges) == 0 {
-		return
-	}
-	var pending []FreebucksPriceChange
-	var ready []FreebucksPriceChange
-	for _, c := range fb.PriceChanges {
-		at, err := time.Parse(time.RFC3339, c.At)
-		if err != nil || at.After(now) {
-			pending = append(pending, c)
-			continue
-		}
-		ready = append(ready, c)
-	}
-	if len(ready) == 0 {
-		return
-	}
-	sort.Slice(ready, func(i, j int) bool {
-		ai, _ := time.Parse(time.RFC3339, ready[i].At)
-		aj, _ := time.Parse(time.RFC3339, ready[j].At)
-		return ai.Before(aj)
-	})
-	if fb.Prices == nil {
-		fb.Prices = map[string]float64{}
-	}
-	if fb.PriceNotices == nil && len(ready) > 0 {
-		fb.PriceNotices = map[string]string{}
-	}
-	for _, c := range ready {
-		if _, ok := fb.Prices[c.ModelID]; !ok {
-			continue
-		}
-		fb.Prices[c.ModelID] = c.Price
-		fb.PriceNotices[c.ModelID] = c.Tagline
-	}
-	fb.PriceChanges = pending
-}
-
-// AvailabilityWindow is the parsed daily availability window from a
-// model_unavailable admission response's availableHours string (issue #158),
-// e.g. "9am ET-5pm PT every day" or "08:00-20:00". Times are normalized to
-// minutes since midnight in US Pacific — the reference zone FreeBuff
-// sessions and quota windows are Pacific-based — so a skip decision needs no
-// DST math at compare time. ET/EST/EDT are converted by subtracting the
-// fixed 3-hour ET→PT offset (both observe US DST in lockstep).
-type AvailabilityWindow struct {
-	// StartMinute/EndMinute bound the daily window, minutes since midnight
-	// Pacific. StartMinute == EndMinute means a degenerate/24-7 window:
-	// callers must not skip on it.
-	StartMinute int
-	EndMinute   int
-	// Raw is the original availableHours string, for logging.
-	Raw string
-}
-
-// availableTimeRE matches "H[:MM] [am|pm] [zone]" - "H[:MM] [am|pm] [zone]"
-// pairs in an availableHours string. Zones are best-effort 2-3 letter
-// tokens (ET/PT/…); unrecognized tails are ignored (the regex is not
-// anchored).
-var availableTimeRE = regexp.MustCompile(`(?i)(\d{1,2})(?::(\d{2}))?\s*(am|pm)?\s*([a-z]{2,3})?\s*-\s*(\d{1,2})(?::(\d{2}))?\s*(am|pm)?\s*([a-z]{2,3})?`)
-
-// ParseAvailabilityWindow parses an upstream availableHours string into a
-// daily window. Supported shapes:
-//
-//	"08:00-20:00"            24-hour window (interpreted in Pacific)
-//	"9am ET-5pm PT every day"  12-hour window with US timezone abbreviations
-//
-// ok is false when no start/end time pair can be extracted, so callers fall
-// back to the plain cache TTL instead of deriving a skip bound.
-func ParseAvailabilityWindow(s string) (AvailabilityWindow, bool) {
-	m := availableTimeRE.FindStringSubmatch(strings.TrimSpace(s))
-	if m == nil {
-		return AvailabilityWindow{}, false
-	}
-	start, ok1 := parseAvailableTime(m[1], m[2], m[3], m[4])
-	end, ok2 := parseAvailableTime(m[5], m[6], m[7], m[8])
-	if !ok1 || !ok2 {
-		return AvailabilityWindow{}, false
-	}
-	return AvailabilityWindow{StartMinute: start, EndMinute: end, Raw: s}, true
-}
-
-// parseAvailableTime converts one time token to minutes since midnight
-// Pacific. hour/minute may be 12-hour (with meridiem) or 24-hour; the zone
-// token converts ET-family zones to Pacific minutes. Unknown/absent zones
-// are interpreted directly as Pacific (best-effort — the caller's TTL cap
-// bounds any misparse).
-func parseAvailableTime(hour, minute, meridiem, zone string) (int, bool) {
-	h, err := strconv.Atoi(hour)
-	if err != nil || h < 0 || h > 23 {
-		return 0, false
-	}
-	min := 0
-	if minute != "" {
-		m, err := strconv.Atoi(minute)
-		if err != nil || m < 0 || m > 59 {
-			return 0, false
-		}
-		min = m
-	}
-	switch strings.ToLower(meridiem) {
-	case "am":
-		if h == 12 {
-			h = 0
-		}
-	case "pm":
-		if h < 12 {
-			h += 12
-		}
-	case "":
-	default:
-		return 0, false
-	}
-	total := h*60 + min
-	switch strings.ToUpper(zone) {
-	case "ET", "EST", "EDT":
-		total -= 3 * 60
-	case "PT", "PST", "PDT", "":
-	default:
-		// Unknown zone: keep as Pacific (best-effort).
-	}
-	total %= 1440
-	if total < 0 {
-		total += 1440
-	}
-	return total, true
-}
-
-// pacificLoc returns America/Los_Angeles, falling back to a fixed PDT zone
-// when the tz database is unavailable (mirrors pool/spend.go's loc helper).
-var pacificLoc = sync.OnceValue(func() *time.Location {
-	if loc, err := time.LoadLocation("America/Los_Angeles"); err == nil {
-		return loc
-	}
-	return time.FixedZone("PDT", -7*60*60)
-})
-
-// AvailableAt reports whether minute-of-day (Pacific) falls inside the
-// window. A degenerate window (StartMinute == EndMinute) is always
-// available: the caller must not skip on it.
-func (w AvailabilityWindow) AvailableAt(now time.Time) bool {
-	if w.StartMinute == w.EndMinute {
-		return true
-	}
-	m := now.In(pacificLoc()).Hour()*60 + now.In(pacificLoc()).Minute()
-	if w.StartMinute < w.EndMinute {
-		return m >= w.StartMinute && m < w.EndMinute
-	}
-	// Overnight window (e.g. 22:00-06:00): open from start until midnight,
-	// then from midnight until end.
-	return m >= w.StartMinute || m < w.EndMinute
-}
-
-// NextStart returns the next wall-clock instant the window opens, strictly
-// after now. For a degenerate window it returns now (no future opening to
-// wait for).
-func (w AvailabilityWindow) NextStart(now time.Time) time.Time {
-	loc := pacificLoc()
-	t := now.In(loc)
-	m := t.Hour()*60 + t.Minute()
-	if w.StartMinute == w.EndMinute {
-		return now
-	}
-	start := time.Date(t.Year(), t.Month(), t.Day(), w.StartMinute/60, w.StartMinute%60, 0, 0, loc)
-	if m < w.StartMinute {
-		return start
-	}
-	// The window already opened today (and we are outside it — a refusal
-	// was cached, so the upstream disagrees with our parse): the next
-	// opening is tomorrow.
-	return start.AddDate(0, 0, 1)
-}
-
-// SessionStanding is the upstream account standing block (issue #96): the
-// pre-join/session response's "standing" field. NextLevelAt is parsed with
-// parseFlexTime; zero when the server omits it.
-//
-// CappedBy/CappedReason name the trust cap holding the account at its level
-// (e.g. third_party_client, anonymous_network — reference/freebuff
-// freebuff-trust.ts FreebuffStandingInfo), Blurb is the human explanation,
-// and NextSteps are the earn-back actions upstream suggests (issue #140).
-type SessionStanding struct {
-	Level        string
-	Label        string
-	Score        float64
-	NextLevelAt  time.Time
-	NextLevel    string
-	CappedBy     string
-	CappedReason string
-	Blurb        string
-	NextSteps    []StandingNextStep
-}
-
-// StandingNextStep is one suggested trust-earning action
-// (FreebuffTrustNextStep, freebuff-trust.ts:415-422).
-type StandingNextStep struct {
-	ID     string
-	Label  string
-	Detail string
-	Points float64
-	Href   string
-}
-
-// ModelQuota is one model's live session quota from the upstream
-// rateLimitsByModel map, per the official CLI wire shape
-// (reference/freebuff/common/src/types/freebuff-session.ts).
-// Entitlement holds the per-period breakdown (base/referral/streak/promo;
-// promo is omitted by default) that sums to Limit when the server emits it.
-type ModelQuota struct {
-	Model       string
-	Limit       float64
-	RecentCount float64
-	ResetAt     time.Time
-	Period      string // "pacific_day" | "pacific_week" | "pacific_month" (empty when absent)
-	// Pool / PoolLabel group models that share one session-quota pool on
-	// the wire (FreebuffSessionRateLimit). Pool is opaque — group by it, never
-	// match on its value; PoolLabel is the server-authored display string.
-	Pool        string
-	PoolLabel   string
-	Entitlement map[string]float64
-}
-
-// rawModelQuota mirrors one rateLimitsByModel entry on the wire. resetAt is
-// parsed with parseFlexTime (RFC3339, unix seconds, or unix ms); windowHours
-// (deprecated) is deliberately not surfaced.
-type rawModelQuota struct {
-	Model                string             `json:"model"`
-	Limit                float64            `json:"limit"`
-	RecentCount          float64            `json:"recentCount"`
-	Period               string             `json:"period"`
-	ResetAt              any                `json:"resetAt"`
-	Pool                 string             `json:"pool"`
-	PoolLabel            string             `json:"poolLabel"`
-	EntitlementBreakdown map[string]float64 `json:"entitlementBreakdown"`
-}
-
-// rawReferral mirrors the session response's "referral" block
-// (FreebuffReferralInfo in reference/common/src/types/freebuff-session.ts).
-type rawReferral struct {
-	Code                    string `json:"code"`
-	ReferrerName            string `json:"referrerName"`
-	QualifiedCount          int    `json:"qualifiedCount"`
-	WeeklySessionsRemaining int    `json:"weeklySessionsRemaining"`
-	ResetAt                 any    `json:"resetAt"`
-	GithubLinked            bool   `json:"githubLinked"`
-}
-
-// rawStanding mirrors the session response's "standing" block (issue #96).
-// nextLevelAt is parsed with parseFlexTime.
-type rawStanding struct {
-	Level        string            `json:"level"`
-	Label        string            `json:"label"`
-	Score        float64           `json:"score"`
-	NextLevelAt  any               `json:"nextLevelAt"`
-	NextLevel    string            `json:"nextLevel"`
-	CappedBy     string            `json:"cappedBy"`
-	CappedReason string            `json:"cappedReason"`
-	Blurb        string            `json:"blurb"`
-	NextSteps    []rawStandingStep `json:"nextSteps"`
-}
-
-// rawStandingStep mirrors one FreebuffTrustNextStep on the wire.
-type rawStandingStep struct {
-	ID     string  `json:"id"`
-	Label  string  `json:"label"`
-	Detail string  `json:"detail"`
-	Points float64 `json:"points"`
-	Href   string  `json:"href"`
-}
-
-type rawFreebucksWindow struct {
-	Limit     float64 `json:"limit"`
-	Spent     float64 `json:"spent"`
-	Remaining float64 `json:"remaining"`
-	ResetAt   any     `json:"resetAt"`
-}
-
-type rawFreebucksWallet struct {
-	Balance      float64 `json:"balance"`
-	MonthlyBonus float64 `json:"monthlyBonus"`
-	NextBonusAt  any     `json:"nextBonusAt"`
-}
-
-type rawFreebucksSpendCeiling struct {
-	LimitUsd float64 `json:"limitUsd"`
-	ResetAt  any     `json:"resetAt"`
-}
-type rawFreebucksMonthlyAllowance struct {
-	LimitUsd     float64 `json:"limitUsd"`
-	SpentUsd     float64 `json:"spentUsd"`
-	RemainingUsd float64 `json:"remainingUsd"`
-	ResetAt      any     `json:"resetAt"`
-}
-
-// rawFreebucks mirrors upstream FreebuffFreebucksInfo (issue #321 wire
-// drift, #350 for exemption/notices/schedule): spendable balance + the
-// daily pool window + the never-expiring wallet + the USD spend ceiling +
-// the monthly allowance + the plan id + the quota exemption + per-model
-// price-notice copy + the announced repricing schedule.
-type rawFreebucks struct {
-	Balance      float64                       `json:"balance"`
-	Daily        rawFreebucksWindow            `json:"daily"`
-	Wallet       *rawFreebucksWallet           `json:"wallet"`
-	Spend        *rawFreebucksSpendCeiling     `json:"spend"`
-	Monthly      *rawFreebucksMonthlyAllowance `json:"monthly"`
-	PlanID       *string                       `json:"planId"`
-	QuotaExempt  *bool                         `json:"quotaExempt"`
-	Prices       map[string]float64            `json:"prices"`
-	PriceNotices map[string]string             `json:"priceNotices"`
-	PriceChanges []rawFreebucksPriceChange     `json:"priceChanges"`
-}
-
-// rawFreebucksPriceChange mirrors FreebuffPriceChange (issue #350).
-type rawFreebucksPriceChange struct {
-	At      string  `json:"at"`
-	ModelID string  `json:"modelId"`
-	Price   float64 `json:"price"`
-	Tagline string  `json:"tagline"`
-}
-
-type rawFreeWindows struct {
-	DayUsed      float64 `json:"dayUsed"`
-	DayLimit     float64 `json:"dayLimit"`
-	WeekUsed     float64 `json:"weekUsed"`
-	WeekLimit    float64 `json:"weekLimit"`
-	MonthUsed    float64 `json:"monthUsed"`
-	MonthLimit   float64 `json:"monthLimit"`
-	DayResetAt   any     `json:"dayResetAt"`
-	MonthResetAt any     `json:"monthResetAt"`
-}
-
-type rawSubscription struct {
-	DayUsed            float64  `json:"dayUsed"`
-	DayLimit           float64  `json:"dayLimit"`
-	FiveDayUsed        float64  `json:"fiveDayUsed"`
-	FiveDayLimit       float64  `json:"fiveDayLimit"`
-	MonthUsed          float64  `json:"monthUsed"`
-	MonthLimit         float64  `json:"monthLimit"`
-	DayPremiumUsed     float64  `json:"dayPremiumUsed"`
-	DayPremiumLimit    float64  `json:"dayPremiumLimit"`
-	DayResetAt         any      `json:"dayResetAt"`
-	PeriodEndsAt       any      `json:"periodEndsAt"`
-	MonthSpendUsd      float64  `json:"monthSpendUsd"`
-	MonthSpendLimitUsd float64  `json:"monthSpendLimitUsd"`
-	FreeDayUsed        *float64 `json:"freeDayUsed"`
-	FreeDayLimit       *float64 `json:"freeDayLimit"`
-}
-
 // parseSessionResponse decodes a session control response body into a
 // SessionState: the 404 create/poll mapping, JSON decode, quota/standing/
 // availability-window parsing, and the passive ban-risk feed (#64). Errors
 // are classified through the standard matrix.
-func windowFromRaw(w rawFreebucksWindow) FreebucksWindow {
-	out := FreebucksWindow{Limit: w.Limit, Spent: w.Spent, Remaining: w.Remaining}
-	if t, err := parseFlexTime(w.ResetAt); err == nil {
-		out.ResetAt = t
-	}
-	return out
-}
 
 func (c *Client) parseSessionResponse(req *http.Request, resp *http.Response, body string) (*SessionState, error) {
